Button logic in feedback dashboard
